requirements.txt
alembic==1.13.2
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
dnspython==2.6.1
eventlet==0.36.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-DebugToolbar==0.15.1
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
flask-profiler==1.8.1
Flask-SQLAlchemy==3.1.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.3.2
pytest-cov==5.0.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==72.2.0
simplejson==3.19.2
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
zope.event==5.0
zope.interface==7.0.1
